Changed to make DispatchTime an interface and the use the factory  pattern to create instances.
Changed to make DispatchTime an interface and the use the factory  pattern to create instances.
