+++ .github/workflows/dotnet.yml
@@ -1,0 +1,26 @@
+name: .NET
+
+on: [ push, pull_request ]
+
+jobs:
+  build-job:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Setup .NET 7.x
+      uses: actions/setup-dotnet@v1
+      with:
+        dotnet-version: 7.x
+    - name: Restore dependencies
+      run: dotnet restore
+    - name: Build
+      run: |
+        dotnet build DispatchR.sln \
+          -c Release
+    - name: Test net7.0
+      run: |
+        dotnet test DispatchR.Test/DispatchR.Test.csproj \
+          -c Release --no-build --no-restore \
+          -f net7.0 \
+          -v normal \
+          --filter "TestCategory!=very.slow"
