+++ DispatchR.Test/DispatchR.Test.csproj
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Dispatchr\DispatchR.csproj" />
+    <ProjectReference Include="..\DispatchR\DispatchR.csproj" />
   </ItemGroup>
 
 </Project>
