+++ DispatchR.Test/MockDispatchee.cs
@@ -3,7 +3,7 @@
     [DispatchOrder(1)]
     public class MockDispatchee1 : Dispatchee
     {
-        public MockDispatchee1(Func<CancellationToken, Task> func, DispatchTime dispatchTime) : base(func, dispatchTime)
+        public MockDispatchee1(Func<CancellationToken, Task> func, IDispatchTime dispatchTime) : base(func, dispatchTime)
         {
             
         }
@@ -12,7 +12,7 @@
     [DispatchOrder(2)]
     public class MockDispatchee2 : Dispatchee
     {
-        public MockDispatchee2(Func<CancellationToken, Task> func, DispatchTime dispatchTime) : base(func, dispatchTime)
+        public MockDispatchee2(Func<CancellationToken, Task> func, IDispatchTime dispatchTime) : base(func, dispatchTime)
         {
 
         }
